imapd: add client_behavior, to track what cliends do
The goal here will be to accumulate "I saw the client use feature X"
during each connection.  We will log those features, along with the
username and client id (the "name" part of ID submision, if any).

Later, we can use this to say "for all log lines in a period of time,
union all the seen features for each client".

This commit only adds a very basic form of this structure, and will be
followed by a very basic test.  Subsequent commits will just add new
feature flags.