upgrade.rst: remove instructions for upgrades before 3.2
The only way to upgrade to 3.6 is to have first version 3.2 installed.
That said, it can be assumed that users will first read the upgrade
instructions to version 3.2 from https://www.cyrusimap.org/3.2/imap/download/upgrade.html .

Since there is stated already that BDB is gone in Cyrus IMAP 3.0 and the
handling of special-use flags after 2.4 is discussed, there is no need to
have this information in the upgrade notes for 3.6, 3.8 and master/dev .

Likewise for the other changes.