client behavior: track when ESEARCH is used
We log this as "did_multisearch" instead of "did_esearch" because (rjbs
thinks) the naming of things has been abused. ;-)

ESEARCH is a capability that extends SEARCH.

MULTISEARCH is a capability that adds the ESEARCH command.

"Multisearch" is not ambiguous, but "esearch" is, so I used
"multisearch".  Also, you can't spell multisearch without "murch".  And
also "its ale", or maybe "I steal".