http_caldav: do not add default alarms to iCalendar on disk
Up to now, whenever a VEVENT was PUT over CalDAV with the
X-JMAP-USEDEFAULTALERTS property being true, we rewrote
whatever current default alarms that user had configured
in the iCalendar data on disk.

But this isn't necessary because we inject default alarms
in the iCalendar data both during (multi)GET and when
reading calendar data in calalarmd.

Keeping the original VALARMs now also allows us to stay closer
to JSCalendar, which also preserves whatever alerts are set
in the "alerts" property, regardless of the "useDefaultAlerts"
property.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>