mboxname: global locking support
if `global_lock` is configured in imapd.conf (default on) then always
take a SHARED lock to "$GLOBAL" before taking any write locks.  This
lock gets inserted into the open_mboxlocks before the first write lock,
and only released when it's the front of the queue, so ANY locks taken
past there also hold it.

It only releases once all the write locks and anything taken after them
are released, guaranteeing that when this is NOT locked, there are no
writes or files in partially changed states.

This means that you can take an exclusive lock on $GLOBAL and know that
no files are in flux - so you can get a clean snapshot while that lock
is held.