auth.h: installed headers can't #include "lib/..."
When our headers are installed, they're no longer buried under a
"lib/" subdirectory, which means installed headers using this spelling
can't find their dependencies, and that means third party code that
wants to link against libcyrus et al cannot be compiled.

Installed headers must spell it `#include "foo.h"`, not `#include "lib/foo.h"`

md5.h is not  installed, so the spelling it uses should not make a difference.
xsha1.h is not installed, but ought to be.  I've fixed the spelling in both
of these for consistency