http_carddav: rewrite vCard 4.0 in CardDAV GET
This works around a bug that had caused Cyrus to write bogus
version 4.0 vCards to disk. The issue was that embedded binary
data in version 4.0 must be encoded as a data: URI, but it
was written as a version 3.0 b-encoded value.
The bug is fixed in b8a879ccf22d52d336662d506d3a14ddf341b60b

This patch changes CardDAV GET to always rewrite a version
4.0 vCard before sending the response to the client, even if
the vCard already is having version 4.0. As a result, clients
get correctly encoded property values for both v3.0 and v4.0.