http_jmap: store uploaded blobs as application/octet-stream
JMAP blobs are untyped and storing the client-supplied media type
may result in Cyrus message blobs having a wrong Content-Type
header. This changes the JMAP blob upload handler to store
blobs as application/octet-stream content, unless the uploaded blob
is of type message/rfc822 and the blob passes our checks for valid
messages of that type.