message.c: support encoded message/rfc822 attachments
Up until now the Content-Transfer-Encoding header for message/rfc822
attachments was ignored. That is fine for rfc822 but we want to treat
message/rfc822 and message/global equally, and the global subtype
does allow its content to be encoded.

This patch changes the parse code to not throw away the encoding
identifier of a message/rfc822 body part before storing it in cache.
It bumps the mailbox minor cache version to indicate this and handles
previous versions on read.