Cassandane: propagate failures in tiny test loader
    unless (eval "package $into; do qq{$test}; 1") {
         Carp::confess("tried to load $test but it failed: $@");
    }

This was the old code for loading each tiny test file.  If the file
didn't exist, it would throw an exception saying that the file could not
be found.

Unhappily, it would *not* do that on a syntax error.  `do FILE` will set
`$@` on failure to compile and execute, but it does not, itself, die.

The new code will throw the error, if any error was found while loading
the file.