cassandane: remove a literal NUL from a test
This is just bizarre!  Why was there a NUL here?  Why does it compile
without issue?  Well, apparently NUL is just ignored in some number of
places in Perl source, and this is news to me.

Anyway, it isn't a load-bearing NUL.  It isn't part of a string.  But it
*did* confuse perltidy, so let's yank it out.