xapian_wrap.cc: normalize text using libicu's NFKC with CaseFolding
This changes the Xapian backend to normalize any text using NFKC with
CaseFolding before indexing and querying. It bumps the Xapian index
version to 17. The default stemmer keeps using lowercase Cyrus Search Form.

If any of the existing indexes contains version 16 or less, then queries
search for any terms that either match the NFKC-normalized term or the
unnormalized term.