index.c: convert all search part text before indexing
Up until now, Cyrus did not convert the contents of header fields
and a couple of iCalendar properties to search form. This required
us to convert those part contents to lower case in the Xapian
wrapper, but we rather want this all proper NFKC-casefolded.

Calling charset_convert on these values before sending them to
the indexer makes sure all indexed text is in proper form.