Config: make smtp configuration invalid by default
Cyrus will default to sending outgoing mail via sendmail if not
otherwise configured, which is potentially a problem when running
Cassandane on a machine configured to send mail to the world.

Make Cassandane set Cyrus up with an invalid smtp configuration by
default, so that if some test case hits smtpclient_open without
having been set up for it, the test just breaks instead of sending
mail.