Makefile.am: 4-space indents, except where we can't
This file is an Automake file, which means we can use 4-space
indents like the rest of our code.  We were already doing this
in some places in this file; now we do it throughout.

The exceptions are the snippets of Makefile syntax, where hard
tabs are required by make.