jmap_ical: do not use icaltime_is_valid_time to validate datetimes
The libical function 'icaltime_is_valid_time' is a misnomer in multiple
aspects: it does not actually validate the time part of a datetime
and it only checks the year part of the date. It arbitrarily requires
years to be less than 3000, whereas RFC 5545 allows years to be in the
range 0 to 9999 (inclusively).

Some iCalendar clients set datetimes in the year 9999 in the ACKNOWLEDGED
property for VALARMs, and while that's kind of silly we must not reject
such dates as invalid.