tiny-tests: Update to prevent '... will not stay shared' warning
tiny-tests/Caldav/options was throwing:

  Variable "$caldav" will not stay shared at ./tiny-tests/Caldav/options line 11.

This is because it was defining a subroutine that was closing over
variables inside of another subroutine. This generally doesn't do what
you expect, so instead lets make the subroutine itself an anonymous
subroutine which will be generated at runtime and close over outer scope
variables properly.

In newer Perl we could use a 'my sub...', but I don't believe we are aiming
to require a high enough version of Perl yet.