tiny-tests: Rework test to actually compare/test what it should be
Back in d7c8b4f8711 the uid property was removed from default alerts.

This test wasn't updated to account for that, but continued to pass because

    my $uid = (values %{$defaultAlertsWithTime})->{uid};
    $self->assert_matches(qr/UID:$uid/, $res->{content});

would end up with an undef value in $uid and so the assertion would succeed.

Now with strict we see that (values %{$defaultAlertsWithTime})->{uid}
crashes because values uses in this way returns a count of how many items
were in the hash, and so we try to deref that as a hashref and fail.

Instead, ensure that depending on which alarms are currently configured,
we get the correct alarm id for both the owner and the sharee.