tiny-tests: Don't overwrite subs from other tests
This is something we need to solve more generally, but the main problem is
that each tiny test actually exists in the parent package that has loaded it,
so any subroutines it defines is visible in the parent.

This means if two tests define the same sub, one of those subs will be
clobbered. Ugh.

For now, let's side step the problem by giving the two (identical)
subroutines different names.

This is a hack, but puts us in a better place, and we can solve the real
problem later.