cyrusdb: add yield API
This API is designed to be used for databases with MVCC support where
you can take a "read lock" view that allows other writes to happen in
parallel while still getting a snapshot read.  We can use this API
call to reduce the number of fcntl calls needed when doing multiple
queries fast, but wanting to still release occasionally to allow
other operations to progress.

In particular, this is a good thing to call - even with a write lock -
when you're doing some other work; so other processes can take a read
from before your transaction started!