Sieve: allow private Final-Recipient in rejects
This replaces a longstanding private commit in Fastmail's Cyrus, which
replaces Final-Recipient with an encrypted version of the original
recipient.  The goal is to provide a string that can be traced back to a
recipient by the Cyrus operator, but doesn't disclose the final delivery
username to the bounce recipient.

Using session_id should accomplish the same result, at least if logs are
maintained.  It's also a lot less complex than shipping our own little
RC4 implementation!