http_caldav_sched.c: fix for organizer getting stomped in sched_request()
When the organizer gets stomped it can result in binary data
being used when building the JSON structure which then results
in an 'Invalid UTF-8 string' error.

The stomping of the organizer value seems consistent if the
iTIP request contains a large number of ATTENDEEs, but
the JSON failure is random depending on what data the
stomped organizer contains.