jmap_ical.c: reject non-integer priority value
The conditional that validates the priority property value to
be in the valid range from 0..9 had an error, effectively
accepting any JSON value as a valid priority value. Any
non-integer JSON value resulted in priority 0 to be set.

This patch keeps accepting any JSON integer value as priority
(contrary to RFC8984 requiring it to be in the 0..9 range),
but now rejects any non-integer JSON value.

Fixes #5226