Makefile.am: enable compile warnings in CFLAGS
This enables the "all" and "extra" warning diagnostics of the
compiler and instructs it to treat any warning as compile error.

Most of the current Cyrus IMAP maintainers have these flags
enabled in their custom build scripts, but we now want to
establish them as baseline for any builds.

Tested with gcc version 12.2.0 and clang version 14.0.6.