sieve_db.c: always set authstate, not just for setting active
When we do a SieveScript/set.create to make a script, plus a
onSuccessActivateScript, we would set the authstate *only* for the
onSuccessActivateScript.  This worked in two parts:

1. append happens, as nobody, which works because of an "anyone p" ACL
2. flag set happens on onSuccessActivateScript, which worked because we
   set the auth state

In writing tests for the Annotator, I found that it couldn't be
instructed to add flags, because it didn't have permission.  That's
because the annotator runs during step 1, append, when it was acting as
nobody.

We don't want the annotator to run, but we also want it (and other
things) to see the right auth state.  So, we set the auth state up
correctly always, now.