message.c: do not use strlen to calculate header section size
This changes the MIME parser to include any illegal NUL bytes
in the header size of a body part. Before, it read those NUL
bytes as part of the header section but set the header size
only to the count of bytes leading up to the NUL byte.

This mismatch in the stored header size led to an error message
indicating that the file size of the message does not match the
sum of its header and content size. This was a red herring and
now is prevented by setting the correct size and offsets.

Note that the MIME parser did not reject MIME message with
NUL bytes, and this patch keeps it at that. This is to keep
parsing such messages when they already are stored on disk.
New messages with NUL bytes already are rejected at delivery.