global.h: use function type taking any arguments for SASL callbacks
The SASL API defines callback functions to be of type

int (*proc)(void);

but then calls these functions with differing argument types.
This is not sane and won't be allowed starting from C23 onwards.
clang-19 now rejects the current way we set these callbacks.
Until the callback functions are declared sanely in the SASL API
we'll cast the callbacks to a function take any arguments.

Thanks to @ksmurchison