Cassandane: If built with asan/ubsan, detect and report errors
Unlike --valgrind, asan and ubsan are baked into the running programs. If
they happen to be compiled in, they'll write to the provided log
directories, and we can pull out any errors they generate automatically.

Note that unfortunately there are some buggy (still?) verisons of clang
and/or gcc where the UBSAN_OPTIONS log_path overrides the ASAN_OPTIONS
log_path. The effect of this is that tests run with ASan will misreport
errors as ubsan errors, but I think we can live with that.