jmap_mail.c: rewrite CR & LF to CRLF in message/rfc822 attachments
Before, Email/set{create} rejected body parts of type "message/rfc822"
if the body value contained a CR or LF octet that is not part of a
CRLF sequence. Now, bare CR or LF get rewritten to CRLF when converting
the Email object to MIME.

This already is the case for "text/plain" attachments. Typically, CRLF
sequences in attachments of these content types got inadvertently
rewritten to bare LF during download or in text editors.