twom: write the initial header directly
This uses file IO to create the initial file rather than mmap, meaning
hopefully that a crash NEVER leaves us with a broken file.  At least it
will be a lot harder for that to happen.  An IO error writing the header
will also cause us to try to truncate the file short AND unlink the
file.

This commit also fixes a bug with version number writing where it was
writing a 32 value into 16 bits of space (found before zeroing out the
scratch pad)