twom: don't repair DIRTY database when readonly
This takes advantage of the way the pointers are set; that any pointer
that isn't into the future will still point to the correct location,
so by using `end` we can actually get a clean read.  Now if there's a
dirty record at level N and we hit it, we'll wind up dropping a level
and progressing at the lower level, so we might drop to level 0 and
have a much more expensive `locate()` than we would otherwise have,
but it's still a clean read!