master.c: Process babysat daemons before regular daemons
The logic for "wait 10 seconds then retry" for daemons that are failing
frequently doesn't seem like it ever worked. The reason is that we'd respawn
children for *any* services missing children before checking babysat
services.

Reverse the order of checks to see if a service is babysat and needs a child
before spawning off children.

This does mean that after 10+ seconds, we'll only spawn one child at first.
Later we'll come along and spawn more.