cassandane: Add a test for restarting babysat services after 10 seconds
The test kills a bunch of lemmings, then makes sure that we don't have a new
lemming for at least 5 seconds. Then it waits for up to 20 more seconds for
a new lemming to appear.

Previously, cyrus would restart children as fast as it could so this test
would fail.