util.c: bin_to_lchex needs an aligned hex buffer
bin_to_lchex casts a char* to a uint16_t*, then uses that pointer to write
2 bytes at a time. This requires an aligned pointer, so we'll use malloc
to give us a scratch space if the incoming hex pointer is unaligned.

malloc is gauranteed to return aligned pointers for any built-in types.

With this, gcc 12.2.0 (Debian 12.2.0-14) with -fsanitize=undefined (UBSan)
runs clean.