twom.c: don't truncate file->size during recovery
This fixes a bug where aborting a twom transaction followed
by closing the database leaves a dangling mmapped region as long
as the process still is running.

The bug is the result of truncating the file->size field in the
recovery1 function. If the file has previously been mmapped to a
larger region than that new size, the munmap following a db_close
leaves the remaining region mapped for the process.