twom: also don't truncate the file!
There's no need to truncate the underlying file - this was just a way
to ensure that everything that had been written in the bogus incomplete
transaction was zeroed out - but it turns out to have been buggy!  It
won't hurt to leave it there, it will get overwritten later anyway.