Remove ALL tabs (again)
Welcome "git blame" user.  Maybe you have also seen 1f19b582295.

In that commit, in mid-2015, we removed all hard tabs, changing the
style guide to "all spaces, 4 spaces per indent level, no tabs".  After
that change, no automation was put in place to prevent the tabs from
coming back… and they did.

This branch re-removes tabs, and adds a GitHub Action to detect new
ones.

It also adds the file .git-blame-ignore-revs, which includes
instructions for configuring your git repo to ignore this and other
whitespace cleanup commits.

    1. get the sha of this commit as $THIS_COMMIT
    2. git checkout <your-branch>
    3. git rebase $THIS_COMMIT^
    4. git filter-branch --tree-filter 'tools/hard-tab-tool --really' $THIS_COMMIT^..HEAD
    5. git rebase $THIS_COMMIT

NOTE the `^` characters above.  It is very important that you keep the ^
so that this commit is the one that collects all the whitespace changes.