cassandane: Add SuppressLSAN magic to ignore known leaks in tests
This magic lets you define a list of patterns (space separated) that end
up in a LeakSanitizer suppression file on a per-test basis. That test is
then run with only its suppressions enabled, and if anything *but*
suppression output is detected in the *san log files, the test fails.

This makes it easier to ignore things until we can fix them later, if ever...