cassandane: replace literal tabs with \t
We want to remove tabs from our source.  This file is being changed
unlike most of the bulk-fixing, as we do need tabs in the data, but we
don't need literal tabs to accomplish it.  Since we can't just run the
"replace tabs with spaces fixer on this file", I've fixed it by hand and
put it in its own commit.