docs: remove a bunch of install-from-distro documentation
Many of these distributions don't have Cyrus anymore, or still have a v2
version of Cyrus.  In the cases that still have them, like Debian, I
don't think the documentation is sufficient, and we should remove it
until we write something that *actually* gets a user from a Debian
prompt to a *usable* Cyrus install.