cyr_buildinfo: Include compilation info about asan/ubsan/gcc
This is necessary because cassandane may need to LD_PRELOAD the correct
asan library when running certain tests, so it needs to know if it's
preloading a gcc sanitizer library or a clang sanitizer library.

Also it may need to load or not load suppressions, etc, if certain
sanitizers are active, and make other decisions.