xunlink: return +1 when nothing removed because ENONET
Recently, xunlink was changed to restore errno after a call if the
result was ENOENT, and to return 0 in that case.  The theory was that we
use xunlink when we know the file might not exist, and that we should
treat ENOENT like success.  "If we care about the result, we should use
unlink, not xunlink!" we said, at the time.

Unsurprisingly, we've now found a case where we care.  During append, we
try to remove any file that might be in our way.  If this works, it
means there *was* a file, and we report that we removed it.
Unfortunately, when xunlink changed, it seemed like there was *always* a
file, because we started to get 0 instead of ENOENT.

When xunlink was changed, we said "any time we care about tracking
ENOENT, we should use unlink, not xunlink".  In fixing this bug, I found
that I disagree.  I think xunlink has three cases, each of which is
interesting in different contexts:

1. a file existed and was unlinked
2. no file existed, nothing unlinked
3. an error other than ENOENT occurred

This change makes xunlink able to indicate which of the three cases
occurred by returning, respectively, 0, +1, or -1.

It also updates every caller of xunlink to test for the correct case.
Anything that was checking for an error, before, now checks for -1.  The
one place that checked for `== 0` has been left intact.  It still really
wants to test for an actual unlink occurring.