docs: big-ish reorganization of release notes
First, divide releases into three parts:

1.  current development
2.  supported versions
3.  older versions

So, first and foremost is "the current development version of Cyrus".
Then come the releases we support.  (I think I've selected the right
list, but it'll be easy to fix if I'm wrong.)  Then old versions that we
won't support.

This also groups the (3.even) releases with the (3.odd) snapshots that
preceded them.  This means the list is more strictly historical.

Secondly, I merged the 2.4-dav betas into the 2.4 series.