cassandane/Makefile: do not run the syntax test by default
We don't actually need to do a syntax test as part of make all.  It can
be its own thing, run on demand.  My thinking is that it should be run
by the human or robot running the tests, based on whether they think
they're likely to *save* or *lose* time by compiling everything up
front.  For a two-test run, `make syntax` is probably a waste.  For all
the tests, it's just noise.  (On my MacBook Air M2, it takes 7s.)

The next step will be to make `cyd test` run `make syntax` iff there are
no positive test specifiers given.  (That is: run it for `cyd test` and
`cyd test !Foo` but not `cyd test Bar`.)