Cass: Make sure we use Error::Simple objects for make_new_from_error
Error.pm requires this, of any exceptions thrown (by, say, JMAP::Tester)
are lost entirely.

It's silly, but Error.pm does:

    $err = $Error::ObjectifyCallback->( { 'text' => $err } ) unless ref($err);

So if you die with an exception object, or a reference to a hash or array,
later code in Error.pm expects an Error::Simple object but instead gets
whatever you gave it, and crashes. Ugh.