slowio: Use CLOCK_PROCESS_CPUTIME_ID instead of CLOCK_MONOTONIC
On systems with slower disks or heavy load, the cyrus processes may not get
much CPU time. This can skew the slowio logic when it's trying to determine
the rate of speed at which data is being read or written.

For example, if the process has had only .0008 seconds of CPU time, but it's
been waiting on disk IO for .004s, and it's processed 16384 bytes, it will
calculate:

    16384 / .004 = 4,096,000 bytes/sec

...but the threshold to delay is 6,250,000 bytes/sec.

The CPU time would have caused a delay since:

    16384 / .0008 = 20,480,000 bytes/sec.

This seems reliable to "fix" the tests on my boxes which have slow disks,