tls.c: new_session_cb() should return 0 rather than 1
This prevents leaking SSL_SESSIONs when caching.

The reasoning for this is found in SSL_CTX_sess_set_new_cb(3):

If the callback returns 0, the session will be immediately removed
from the internal cache and the reference count released.
If the callback returns 1, the application retains the reference
(for an entry in the application-maintained "external session cache"),
and is responsible for calling SSL_SESSION_free() when the session
reference is no longer in use.