configure.ac: make 'function nesting' check safer
The call to `func("abc", ...)`, if run, would have qsorted the
bytes of the string literal "abc", which is undefined behaviour.

When building with -Wwrite-strings and -Werror (as we usually do),
the attempt to write to a string literal is a compile-time error.
The check misinterprets this as function nesting being unavailable.