logging: log good and bad logins with logfmt
This is the first big push toward logfmt-based logging, chosen not
because it's the most important thing to convert, but because it was a
fairly wide-ranging update with a limited subject matter.

One next step is to write down what a "login.good" and "login.bad" event
should look like and stick it in notes.  Even without that, this makes
logins easy to machine-parse.

Special note, the loss of `popd_apop_chal` in popd login failure
logging.  I think this is acceptable, largely because I don't know what
anybody would want to do with the challenge in their logs!