Cass: Fixup encode_base64jmap and add decode_base64jmap and tests
encode_base64jmap failed to encode '-' since it was being treated as a range
character in the tr// rather than a literal, so we must escape it first with
'\'.

This caused heisenfails in Cyrus::Conversations.move_200 when the ids would
contain a '-' on occasion.

Add tests, and also add a related decoder incase tests need it.