Cass: Add test showing Cyrus can't read broken v19 mailboxes
These mailboxes were created when a buggy deployed cyrus rewrote
cyrus.index files with the wrong RecordSize (144 insstead of 112) for
v19 records.

Newer Cyrus asserts that the RecordSize is correct and refused to read these
files in so they cannot be fixed by a reconstruct.

This test creates such a bad file.