jmap_mail.c: use safe string comparison for body part types
The MIME parser in almost all cases guarantees the body part
media type to be set, but it does not for invalid MIME messages
that already are on disk. This is a known issue that needs to
get fixed in "reconstruct" first.

Also see https://github.com/cyrusimap/cyrus-imapd/pull/5254