ctl_cyrusdb.c: log failure to open at a non-debug level
When we can't open a mailbox for which we have an mbentry, that is
weird, and we should actually log it.  Today it seems we hit this when a
replica had long-deleted folders so old that the master no longer had
tombstones.  This log line would've gotten us to the solution just a bit
faster, and should generally never happen.