calsched_support.c: remove <pref> from from calendar-user-address-set
This changes the value format for the calendar-user-address-set CalDAV
annotation. It reverts the format back to its version before the
implementation of the defaultParticipantIdentityId of the (now obsolete)
JMAP CalendarPreferences object.

The now reintroduced format is to store the list of calendar user
addresses as a comma-separated list of encoded URI values
("mailto: URIs for all practical purposes). The previous value format
preceeded that list with a numeric field, <pref>, followed by a
semi-colon. The <pref> field indicated the position of the
defaultParticipantIdentityId in the ordered list of calendar user
addresses. This was to allow keeping the default participant identity
for JMAP Calendars and the default calendar user address for CalDAV
separate.

Keeping JMAP Calendars default participant identities and CalDAV
calendar user addresses separate was a mistake and this patch fixes
that. As a consequence, setting the defaultParticipantIdentity to
null isn't possible anymore and patching the DAV property with
PROPPATCH also changes the defaultParticipantIdentity in the JMAP
CalendarPreferences singleton.