http_caldav.c: start and end calendar-user-address-set with default href
This changes the PROPFIND return value for calendar-user-address-set.
In addition to returning the default calendar user address at the
*end* of the list of hrefs, this patch also sets a XML href node at
the *start* of the list, effectively duplicating the preferred href
in the value.

CalDAV clients interpret the value of calendar-user-address-set
differently and Thunderbird recently changed its logic from picking
the *last* entry to picking the *first*. This patch is meant to
allow interoperating with both older and newer Thunderbird clients,
as well as Apple and others, presuming that clients ignore any
href other than the one they are hard-coded to pick in the list.