charset.c: discard empty escape sequences from ISO-2022-JP text
This preprocesses text encoded in ISO-2022-JP before handing it
over to libicu for conversion. The character encoding is defined
in RFC 1468 and contains segments of both one-byte and two-byte
character sequences. A segment starts with an escape sequence that
indicates the encoding of the following text bytes.

Escape sequences must be followed by at least one text character,
but some MIME messages violate this and instead concatenate
multiple escape sequences. For every such sequence, libicu inserts
a REPLACEMENT character. This is bad user experience and other
character encoding libraries, such as the one for Java, Go or the
iconv utility ignore such empty sequences.

This patch tweaks the charset pipeline by discarding all but the
last in a run of escape sequences before passing the bytes to the
libicu-based charset converters. As a result, the converted text
does not contain REPLACEMENT characters, as if those empty escape
sequences never were part of the byte stream.