jmap: don't try to create mailbox for read-only sieve DB activities
Trying to create a mailbox requires a userns write lock, but for /get
and friends, we want to be able to rely on taking just a read_lock,
and hence never blocking any other reads.  So we can just return a
synthetic "nothing here" response if the mailbox doesn't exist.