use 0 rather than UTIME_OMIT for creating timestamps
This is because the utimes functions to actually change the timestamp
on disk will ignore the timestamp if it has UTIME_OMIT, but v19 writes
would have used those.

Since 0 has the same actual meaning now (not UTIME_SAFE_NSEC) we should
just use zero as our flag value for "not set" on the one timestamp
that matters (internaldate) and let conversatioons_adjust_internaldate
fix it.