Cassandane: put a "description" file in most instance base dirs
After a test run, sometimes I want to very quickly find the main
instance for the Foo.bar_baz test.  This is a bit of a pain.

This commit should make it a bit less of a pain.  When the test
represented by &Cassandane::Cyrus::Foo::test_bar_baz is run, the
/tmp/cass/XXXXX directory for its main instance will now have a
`description` file that contains "Foo.bar_baz".

We can improve upon this in the future, but this should help to start.