message.c: do not match conversation subject if $memo keyword is set
This patch disables checking the conversation subject if the "$memo"
keyword is set on a message. It never makes sense to start a new
conversation for a memo, it always belongs to the conversation on which
it got set. This already worked in the general case, but in particular
calendar invitation emails can change their subjects and attaching memos
to their conversation could fail.