duplicate.c: when auditlogging, use clearer field name
We were reporting mailbox uniqueid as "user=<X>", which is much more
opaque than "uniqueid" which we use elsewhere.

Unfortunately, the field isn't just uniqueid.  It's the uuid of the
mailbox in lmtp scope, but in Sieve or (ha!) nntp, it's another kind of
context.  In either case, "to" is a bad name.  So this just changes the
logs so that an operator can realize that if it's followed by a guid,
it's probably a mailboxid.