search_query.c: require top-level message for header queries
A result for a query including a header search criteria might be
a false positive if it matches a message that is both a top-level
Cyrus message and a message/rfc822 attachment in another message.

This patch excludes search results having part ids if the clause
requires at least one header criteria to match.