cassandane: get uniqueid, not mailboxid, to use as uuid
These tests were fetching the MAILBOXID status response, but then using
it in places that UNIQUEID would be more appropriate.  They'd look for
located-by-uuid folder on disk, or $FOLDER_IDS in the conversations db.
If mailbox ids are compact ids, these cases are never going to work.  We
should get the "UNIQUEID" status response instead.