jmap_api.c, jmap_contact.c: implement applyEmptyUpdates for cards
Currently, if a /set{update} request has empty content,
_card_set_update() will short-circuit its processing and simply
bump the modseq of the existing resource without rewriting the
file on disk.

Now, if the empty /set{update} request includes
applyEmptyUpdates:true, the resource will also be rewritten
with updated headers, but without changing the vCard content
(no vCard -> JSContact -> vCard conversion).

This option can be used to update existing contacts to include
their compact JMAP ID in the Content-Disposition header
so that it can be recovered if both the cyrus.index record
and DAV DB records for a resource are lost.  The JMAP ID is
generated from the createdmodseq of the resource, so if we
lose both sources of the createdmodseq, the C-D header is
the last line of defense against losing the JMAP ID, thus
preventing a client from having to completely resync.