libconfig: import dir_hash_c, dir_hash_b from util
The goal is to have these somewhere that they're accessible from
an installed header (i.e. not "util.h"!)

libconfig.h isn't an amazing place for these functions, but it
is an installed header.  And to use these functions you'll need
libconfig.h anyway, because the correct value for the "full"
argument comes from IMAPOPT_FULLDIRHASH.