Add custom includeCADs argument to Email/changes
Normally, if a record has been both created AND destroyed since the
oldState of the client it is omitted entirely from the /changes results.

Extra argument to Email/changes — includeCADs: Boolean (default: false).

If true, then any Email that was both created AND destroyed
since oldState is included in the destroyed array.