cunit: loosen meta-timeofday:time_slowdown and slowio:limit2 tolerances
With Linux on Apple hardware (whether via Docker or Parallels),
the various sleep syscalls seem to reliably sleep a little
longer than was actually requested, causing these tests to fail.

The exact durations in these tests aren't critical, the point is
to check that things are slowed down, which they are.  So just
massage the tolerances until the tests pass.