index_file: add new data types and use them
This adds:
't' for 32bit timestamps
'Q' for 64bit aligned (quad) values
'<' for 32bit disk values unpacked into 64bit memory

And changes lots of places to use them.

This means that the type description itself is enough to describe the full operation,
no need to add ondisk size checks; because it always maps a size to a size.

as part of this we set all struct index_record fields to the exact type we're unpacking into,
and ensure they're aligned well in memory.