message.c: refactor header section parser for clarity
This refactors the code that detects start and end of
header sections for both the top-level message and its
its body parts. The former code was a multi-line if
statement and its logic was not immediately obvious.

This patch splits that if-statement into multiple smaller
conditionals, each described by a comment. This is
meant to improve readability. The original logic is
preserved.